refactor: LoginManager 이름 변경
- User 모델과 관련된 역할을 담당하는 이름으로 변경
